# src/ui/openapi.yaml
openapi: 3.0.3
info:
  version: 1.3.42
  description: |
    REST API surface served by warest-whatsapp-rest-api-multisessions.
  contact:
    name: WARest
servers:
  - url: http://localhost:4001
    description: Local development
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/messages/{messageId}/action/star:
    post:
      operationId: postMessagesMessageIdActionStar
      tags:
        - Messages - Actions
      summary: Star a message
      description: Star a message.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
      responses:
        <<: *commonResponses
        200:
          description: Message starred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Message starred"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "message starred success"
  /api/v1/messages/{messageId}/action/unstar:
    post:
      operationId: postMessagesMessageIdActionUnstar
      tags:
        - Messages - Actions
      summary: Unstar a message
      description: Remove star from a message.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
      responses:
        <<: *commonResponses
        200:
          description: Message unstarred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Message unstarred"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "message unstarred success"
  /api/v1/messages/{messageId}/action/reaction:
    post:
      operationId: postMessagesMessageIdActionReaction
      tags:
        - Messages - Actions
      summary: Add a reaction
      description: React to a message with an emoji.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - emoji
                  properties:
                    emoji:
                      type: string
                      description: Emoji string to apply.
      responses:
        <<: *commonResponses
        200:
          description: Reaction applied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Reaction applied"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "reaction applied success"
  /api/v1/messages/{messageId}/action/unreaction:
    post:
      operationId: postMessagesMessageIdActionUnreaction
      tags:
        - Messages - Actions
      summary: Remove a reaction
      description: Remove reaction from a message.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
      responses:
        <<: *commonResponses
        200:
          description: Reaction removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Reaction removed"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "reaction removed success"
  /api/v1/messages/{messageId}/action/delete:
    delete:
      operationId: deleteMessagesMessageIdActionDelete
      tags:
        - Messages - Actions
      summary: Delete a message
      description: Delete message for current account.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    withMedia:
                      type: boolean
                      default: true
                      description: Delete the media blob when true.
      responses:
        <<: *commonResponses
        200:
          description: Message deleted for me.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Message deleted for me"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "message deleted success"
  /api/v1/messages/{messageId}/action/revoke:
    post:
      operationId: postMessagesMessageIdActionRevoke
      tags:
        - Messages - Actions
      summary: Revoke a message
      description: Revoke message for everyone.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  properties:
                    deleteForMe:
                      type: boolean
                      default: false
                      description: Optionally delete for current account as well.
      responses:
        <<: *commonResponses
        200:
          description: Message revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Message revoked"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "message revoked success"
  /api/v1/messages/{messageId}/action/edit:
    post:
      operationId: postMessagesMessageIdActionEdit
      tags:
        - Messages - Actions
      summary: Edit a message
      description: Edit a previously sent text message.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
                - type: object
                  required:
                    - message
                  properties:
                    message:
                      type: string
                      description: Replacement text.
      responses:
        <<: *commonResponses
        200:
          description: Message edited.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Message edited"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "message edited success"
  /api/v1/messages/{messageId}/action/mark-as-read:
    post:
      operationId: postMessagesMessageIdActionMarkAsRead
      tags:
        - Messages - Actions
      summary: Mark a message as read
      description: Mark a specific message as read.
      parameters:
        - $ref: "#/components/parameters/MessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DestinationSelector"
      responses:
        <<: *commonResponses
        200:
          description: Message read receipts emitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageActionResponse"
              examples:
                success:
                  value:
                    status: true
                    code: 1000
                    message: "Message marked as read"
                    results:
                      - sessionId: "main"
                        phone: "628111111111"
                        to: "62812xxxxxxx"
                        toJid: "62812xxxxxxx@s.whatsapp.net"
                        messageId: "3EB01234"
                        status: "message read success"
...
components:
  parameters:
...
    MessageId:
      in: path
      name: messageId
      required: true
      schema:
        type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-WAREST-API-KEY
      description: "Send your X-WAREST-API-KEY header; use an admin-scoped key for admin-only routes."
  schemas:
    ApiEnvelope:
      type: object
      properties:
        status:
          type: boolean
        code:
          type: integer
        message:
          type: string
        results:
          type: array
          nullable: true
          items:
            type: object
      required: [status, code, message]
...
    SessionSelector:
      description: Provide the session handle by supplying `sessionId`. This identifier is returned by `/api/v1/session/create`.
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string
          description: Unique session identifier returned by `/api/v1/session/create`.
          example: "main"
    DestinationSelector:
      allOf:
        - $ref: "#/components/schemas/SessionSelector"
        - type: object
          required:
            - to
          properties:
            to:
              oneOf:
                - type: string
                  description: Digit string, LID, or JID. Comma separated lists are allowed.
                  example: "6281234567890"
                - type: array
                  items:
                    type: string
...
    MessageActionResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/MessageActionResult"
      example:
        status: true
        code: 1000
        message: Message updated
        results:
          - sessionId: main
            phone: "628111111111"
            to: "62812xxxxxxx"
            toJid: "62812xxxxxxx@s.whatsapp.net"
            messageId: "3EB01234"
            status: "message edited success"
...